Creates protocols for movie list screen
